Eigs and AB model
I changed the eigenvalues function to calculate the smallest eigenvalues instead of the largest. We can keep the largest if you think it's better to, but the smallest are def needed for the spectral gap. I also imported the AB model and invoked the generator as a comment in main. I checked through the GCC function, looks good to me. I'll work on the rest in a bit.