Spectral gap
I'm not sure if it's very scientific to use the number of connected components to determine the spectral gap, so u can ignore that part if u want. For me adding .real in the kdeplot was necessary to avoid errors but if it works fine for u then dw about it.